# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api covid19 Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin6@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get List Of authors",
                "description": "Returns all authors",
                "operationId": "getAllAuthors",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Add a new author",
                "description": "Returns infos on the new added author.",
                "operationId": "AddAuthor",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/authors/id": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get the author using it's id",
                "description": "Returns the author. author must be the author_slug from /authors.",
                "operationId": "getOneAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "delete an author using it's id",
                "description": "Returns nothing . author must be the author_slug from /authors.",
                "operationId": "deleteOneAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "update an existent author",
                "description": "Returns the modified author. author must be the author_slug from /authors/id.",
                "operationId": "UpdateAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get List Of Books",
                "description": "Returns all Books and associated authors and genres. No slug is used during this request",
                "operationId": "getAllBook",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Add a new Book",
                "description": "Returns infos on the new added book with the author and genres associated.",
                "operationId": "AddBook",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/books/id": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get the book using it's id",
                "description": "Returns the book. Book must be the book_slug from /books.",
                "operationId": "getOneBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "delete a book using it's id",
                "description": "Returns nothing . Book must be the book_slug from /books.",
                "operationId": "deleteOneBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "update an existent book",
                "description": "Returns the modified book. Book must be the book_slug from /books/id.",
                "operationId": "UpdateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "path",
                        "description": "title ",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "description": "description",
                        "required": true
                    },
                    {
                        "name": "publication_year",
                        "in": "path",
                        "description": "publication year",
                        "required": true
                    },
                    {
                        "name": "pages_nb",
                        "in": "path",
                        "description": "numbre of pages",
                        "required": true
                    },
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "author's id",
                        "required": true
                    },
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "genre's id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}